sr(api): move UserNotAuthorizedToFindTrainings Error in devcomp
